+++ src/servers/ServersListGroup.tsx
@@ -2,7 +2,7 @@ import { faChevronRight as chevronIcon } from '@fortawesome/free-solid-svg-icons
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { clsx } from 'clsx';
 import type { FC, PropsWithChildren } from 'react';
-import { Link } from 'react-router-dom';
+import { Link } from 'react-router';
 import { ListGroup, ListGroupItem } from 'reactstrap';
 import type { ServerWithId } from './data';
 import './ServersListGroup.scss';